feat(extension): adjust staking view components structure to consume new staking changes
