+++ apps/browser-extension-wallet/src/views/browser-view/features/staking/components/MultiDelegationStaking.tsx
@@ -1,0 +1,8 @@
+import { Staking } from '@lace/staking';
+import React from 'react';
+import { useTheme } from '@providers';
+
+export const MultiDelegationStaking = () => {
+  const { theme } = useTheme();
+  return <Staking theme={theme.name} />;
+};
