+++ apps/browser-extension-wallet/src/views/browser-view/features/staking/components/StakingContainer.tsx
@@ -1,12 +1,17 @@
 /* eslint-disable @typescript-eslint/no-var-requires */
 import { Layout } from '@src/views/browser-view/components';
 import React from 'react';
+import { StakingSkeleton } from './StakingSkeleton';
 
 const Staking =
-  process.env.USE_MULTI_DELEGATION_STAKING === 'true' ? require('@lace/staking').Staking : require('./Staking').Staking;
+  process.env.USE_MULTI_DELEGATION_STAKING === 'true'
+    ? require('./MultiDelegationStaking').MultiDelegationStaking
+    : require('./Staking').Staking;
 
 export const StakingContainer = (): React.ReactElement => (
   <Layout>
-    <Staking />
+    <StakingSkeleton>
+      <Staking />
+    </StakingSkeleton>
   </Layout>
 );
